Added sorting for collection, sorting will be used in profile. Added sets to card display

+++ content/Nenniltoz/Users/views.py
@@ -51,7 +51,7 @@ def login_page(request):
         if user is not None:
             # correct username and password login the user
             auth.login(request, user)
-            return redirect('user_profile')
+            return redirect('user_profile', userID=str(request.user.id))
 
         else:
             messages.error(request, 'Error wrong username/password')
@@ -87,7 +87,7 @@ def register(request):
     logger.debug("Run: register; Params: " + json.dumps(request.GET.dict()))
 
     if request.user.is_authenticated:
-        return redirect('user_profile', username=request.user.username)
+        return redirect('user_profile', userID=str(request.user.id))
 
     if request.method == 'POST':
         f = CreateUserForm(request.POST)
@@ -95,7 +95,7 @@ def register(request):
             f.save(request)
 
             messages.success(request, 'Account created successfully')
-            return redirect('user_profile')
+            return redirect('user_profile', userID=str(request.user.id))
 
     else:
         f = UserCreationForm()
@@ -140,6 +140,7 @@ def build_friend_list(user):
 
     return friend_user_list
 
+
 def build_follower_list(user):
     """Build followers list.
 
@@ -159,49 +160,33 @@ def build_follower_list(user):
     return follower_user_list
 
 
-def player_profile(request):
-    """Display the profile of a user that is not logged in.
-
-    Uses the GET data from request to display user data. Displayed user data is not the current user.
-
-    :param request: GET data: User id/username which is being displayed
-    
-    :todo: Alter to work with non current user
-    """
-    logger.debug("Run: player_profile; Params: " + json.dumps(request.GET.dict()))
-    user = get_object_or_404(User, id=request.user.id)
-    user_profile_obj = UserProfile.objects.get(user=user)
-    friend_obj = build_friend_list(user)
-    pending_obj = build_pending_list(user)
-    follower_obj = build_follower_list(user)
-
-    return render(request, 'Users/UserProfile.html', {'user_profile_obj': user_profile_obj, 'user': user,
-                                                      'friend_obj': friend_obj, 'pending_obj': pending_obj, 'follower_obj':follower_obj})
-
-
 @login_required
-def user_profile(request):
-    """Display the profile of a user that is logged in.
+def user_profile(request, userID):
+    """Display the profile of a user.
 
-    Uses the GET data from request to display user data. Displayed user data is the current user.
+    Uses the GET data from request to display user data.
 
     :param request: GET data: User id which is being displayed
+    :param userID: User id of displayed profile.
 
-    :todo: Alter to work with non current user
+    :todo: None
     """
     logger.debug("Run: user_profile; Params: " + json.dumps(request.GET.dict()))
-    user = get_object_or_404(User, id=request.user.id)
+    user = get_object_or_404(User, id=userID)
     user_profile_obj = UserProfile.objects.get(user=user)
     friend_obj = build_friend_list(user)
     pending_obj = build_pending_list(user)
     follower_obj = build_follower_list(user)
 
+    o_player = not str(request.user.id) == userID
+
     return render(request, 'Users/UserProfile.html', {'user_profile_obj': user_profile_obj, 'user': user,
-                                                      'friend_obj': friend_obj, 'pending_obj': pending_obj, 'follower_obj':follower_obj})
+                                                      'friend_obj': friend_obj, 'pending_obj': pending_obj,
+                                                      'follower_obj': follower_obj, 'o_player': o_player})
 
 
 @login_required
-def send_friend_request(request):
+def send_friend_request(request, userID):
     """Sends friend request
 
     Uses POST data to send friend request to user. Error checks for
@@ -215,9 +200,10 @@ def send_friend_request(request):
     :todo: None
     """
     user = request.POST['curUser']
-    friend=request.POST['newFriend']
+    friend = request.POST['newFriend']
     user_obj = User.objects.get(username=user)
     friend_obj = User.objects.get(username=friend)
+    friend_obj_exist = User.objects.filter(username=friend).exists()
     pending_friends_exist = PendingFriends.objects.filter(user_one=user_obj, user_two=friend_obj).exists()
     friends_exist = Friends.objects.filter(user_one=user_obj, user_two=friend_obj).exists()
     pending_sent_exist = PendingFriends.objects.filter(user_one=friend_obj, user_two=user_obj).exists()
@@ -229,22 +215,21 @@ def send_friend_request(request):
     elif pending_sent_exist:
         messages.success(request, 'Friend request accepted.')
         add_friend(request)
+    elif not friend_obj_exist:
+        messages.success(request, 'Username does not exist.')
     else:
-        try:
-            PendingFriends.objects.create(
-                user_one=user_obj,
-                user_two=friend_obj,
-                created_on=datetime.now()
-            )
-            messages.success(request, 'Friend request sent.')
-        except ValueError:
-            messages.error(request, 'Username does not exist.')
+        PendingFriends.objects.create(
+            user_one=user_obj,
+            user_two=friend_obj,
+            created_on=datetime.now()
+        )
+        messages.success(request, 'Friend request sent.')
 
-    return redirect('user_profile')
+    return redirect('user_profile', userID=str(request.user.id))
 
 
 @login_required
-def add_friend(request):
+def add_friend(request, userID):
     """Creates freind relationship.
 
     Creates the 2 way friend relationship in the database. Removes the pending friend request from the database
@@ -254,7 +239,7 @@ def add_friend(request):
     :todo: None
     """
     user = request.POST['curUser']
-    friend=request.POST['newFriend']
+    friend = request.POST['newFriend']
     user_obj = User.objects.get(username=user)
     friend_obj = User.objects.get(username=friend)
     Friends.objects.create(
@@ -271,11 +256,11 @@ def add_friend(request):
         PendingFriends.objects.get(user_one=friend_obj, user_two=user_obj).delete()
     except:
         PendingFriends.objects.get(user_one=user_obj, user_two=friend_obj).delete()
-    return redirect('user_profile')
+    return redirect('user_profile', userID=str(request.user.id))
 
 
 @login_required
-def process_friend(request):
+def process_friend(request, userID):
     """Processes button click for friend request on receivers profile.
 
     Determines if the user selected to accept or reject the friend request.
@@ -285,21 +270,24 @@ def process_friend(request):
     :todo: None
     """
     user = request.POST['curUser']
-    friend=request.POST['newFriend']
+    friend = request.POST['newFriend']
     user_obj = User.objects.get(username=user)
     friend_obj = User.objects.get(username=friend)
 
-    if 'pendAcceptBtn' not in request.POST:
-        PendingFriends.objects.filter(user_one=friend_obj, user_two=user_obj).update(rejected=True)
-        messages.error(request, 'Friend request rejected. Send new request to add as friend.')
+    if 'userRedirect' not in request.POST:
+        if 'pendAcceptBtn' not in request.POST:
+            PendingFriends.objects.filter(user_one=friend_obj, user_two=user_obj).update(rejected=True)
+            messages.error(request, 'Friend request rejected. Send new request to add as friend.')
+        else:
+            messages.success(request, 'Friend request accepted.')
+            add_friend(request)
+        return redirect('user_profile', userID=str(request.user.id))
     else:
-        messages.success(request, 'Friend request accepted.')
-        add_friend(request)
-    return redirect('user_profile')
+        return redirect('user_profile', userID=str(friend_obj.id))
 
 
 @login_required
-def remove_friend(request):
+def remove_friend(request, userID):
     """Removes a friend
 
     Removes the 2 way relationship for friends.
@@ -309,19 +297,22 @@ def remove_friend(request):
     :todo: None
     """
     user = request.POST['curUser']
-    friend=request.POST['newFriend']
+    friend = request.POST['newFriend']
     user_obj = User.objects.get(username=user).id
     friend_obj = User.objects.get(username=friend).id
+    if 'userRedirect' not in request.POST:
+        Friends.objects.get(user_one=user_obj, user_two=friend_obj).delete()
+        Friends.objects.get(user_one=friend_obj, user_two=user_obj).delete()
 
-    Friends.objects.get(user_one=user_obj, user_two=friend_obj).delete()
-    Friends.objects.get(user_one=friend_obj, user_two=user_obj).delete()
+        messages.error(request, 'Removed friend. Send new request to add as friend.')
 
-    messages.error(request, 'Removed friend. Send new request to add as friend.')
+        return redirect('user_profile', userID=str(request.user.id))
+    else:
+        return redirect('user_profile', userID=str(friend_obj))
 
-    return redirect('user_profile')
 
 @login_required
-def add_follower(request):
+def add_follower(request, userID):
     """Adds a followed user to the user.
 
     Add the followed user to the database.  Error checks for
@@ -333,30 +324,30 @@ def add_follower(request):
     :todo: None
     """
     user = request.POST['curUser']
-    follower=request.POST['newFollower']
+    follower = request.POST['newFollower']
     user_obj = User.objects.get(username=user)
     follower_obj = User.objects.get(username=follower)
+    follower_obj_exist = User.objects.filter(username=follower).exists()
 
     follower_exist = Followers.objects.filter(user_one=user_obj, user_two=follower_obj).exists()
 
     if follower_exist:
         messages.error(request, 'Already following.')
+    elif not follower_obj_exist:
+        messages.error(request, 'Username does not exist.')
     else:
-        try:
-            Followers.objects.create(
-                user_one=user_obj,
-                user_two=follower_obj,
-                created_on=datetime.now()
-            )
-            messages.success(request, 'Follower Added.')
-        except ValueError:
-            messages.error(request, 'Username does not exist.')
+        Followers.objects.create(
+            user_one=user_obj,
+            user_two=follower_obj,
+            created_on=datetime.now()
+        )
+        messages.success(request, 'Follower Added.')
 
+    return redirect('user_profile', userID=str(request.user.id))
 
-    return redirect('user_profile')
 
 @login_required
-def remove_follower(request):
+def remove_follower(request, userID):
     """Removes a followed user
 
     Removes the followed user from the relationship with the current user.
@@ -370,8 +361,9 @@ def remove_follower(request):
     user_obj = User.objects.get(username=user).id
     follower_obj = User.objects.get(username=follower).id
 
-    Followers.objects.get(user_one=user_obj, user_two=follower_obj).delete()
-
-    messages.error(request, 'Removed follower.')
-
-    return redirect('user_profile')
+    if 'userRedirect' not in request.POST:
+        Followers.objects.get(user_one=user_obj, user_two=follower_obj).delete()
+        messages.error(request, 'Removed follower.')
+        return redirect('user_profile', userID=str(request.user.id))
+    else:
+        return redirect('user_profile', userID=str(follower_obj))
